+++ kubeflow/trainer/backends/kubernetes/types.py
@@ -1,0 +1,29 @@
+# Copyright 2025 The Kubeflow Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from typing import Optional
+from kubernetes import client
+
+from kubeflow.trainer.types import types
+
+
+
+class KubernetesBackendConfig(types.BackendConfig):
+    namespace: Optional[str] = None
+    config_file: Optional[str] = None
+    context: Optional[str] = None
+    client_configuration: Optional[client.Configuration] = None
+
+    class Config:
+        arbitrary_types_allowed = True
